feat(search): Some cleanup after search feature.
